+++ pkg/server/server_authz_test.go
@@ -18,6 +18,7 @@ import (
 
 	"github.com/openfga/openfga/internal/authz"
 	"github.com/openfga/openfga/internal/mocks"
+	"github.com/openfga/openfga/internal/utils/apimethod"
 	"github.com/openfga/openfga/pkg/authclaims"
 	"github.com/openfga/openfga/pkg/storage/memory"
 	"github.com/openfga/openfga/pkg/tuple"
@@ -60,6 +61,7 @@ const (
 			define system: [system]
 			define creator: [application]
 			define can_call_delete_store: [application] or admin
+			define can_call_update_store: [application] or admin
 			define can_call_get_store: [application] or admin
 			define can_call_check: [application] or reader
 			define can_call_expand: [application] or reader
@@ -724,7 +726,7 @@ func TestCheckAuthz(t *testing.T) {
 
 		storeID := ulid.Make().String()
 
-		err := openfga.checkAuthz(context.Background(), storeID, authz.Check)
+		err := openfga.checkAuthz(context.Background(), storeID, apimethod.Check)
 		require.NoError(t, err)
 	})
 
@@ -742,19 +744,19 @@ func TestCheckAuthz(t *testing.T) {
 		t.Run("with_SkipAuthzCheckFromContext_set", func(t *testing.T) {
 			ctx := authclaims.ContextWithSkipAuthzCheck(context.Background(), true)
 
-			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
+			err := openfga.checkAuthz(ctx, settings.testData.id, apimethod.Check)
 			require.NoError(t, err)
 		})
 
 		t.Run("error_with_no_client_id_found", func(t *testing.T) {
-			err := openfga.checkAuthz(context.Background(), settings.testData.id, authz.Check)
+			err := openfga.checkAuthz(context.Background(), settings.testData.id, apimethod.Check)
 
 			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
 
 		t.Run("error_with_empty_client_id", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: ""})
-			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
+			err := openfga.checkAuthz(ctx, settings.testData.id, apimethod.Check)
 
 			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
@@ -768,7 +770,7 @@ func TestCheckAuthz(t *testing.T) {
 
 		t.Run("error_check_when_not_authorized", func(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
-			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
+			err := openfga.checkAuthz(ctx, settings.testData.id, apimethod.Check)
 
 			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
 		})
@@ -777,7 +779,7 @@ func TestCheckAuthz(t *testing.T) {
 			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
 			settings.addAuthForRelation(ctx, t, authz.CanCallCheck)
 
-			err := openfga.checkAuthz(ctx, settings.testData.id, authz.Check)
+			err := openfga.checkAuthz(ctx, settings.testData.id, apimethod.Check)
 			require.NoError(t, err)
 		})
 	})
@@ -1626,6 +1628,65 @@ func TestCreateStore(t *testing.T) {
 
 			require.NoError(t, err)
 			require.Equal(t, name, readChangesResponse.GetName())
+		})
+	})
+}
+
+func TestUpdateStore(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+	ds := memory.New()
+	t.Cleanup(ds.Close)
+
+	t.Run("UpdateStore_no_authz_should_succeed", func(t *testing.T) {
+		openfga := MustNewServerWithOpts(
+			WithDatastore(ds),
+		)
+		t.Cleanup(openfga.Close)
+
+		clientID := "validclientid"
+
+		settings := newSetupAuthzModelAndTuples(t, openfga, clientID)
+
+		_, err := openfga.UpdateStore(context.Background(), &openfgav1.UpdateStoreRequest{
+			StoreId: settings.testData.id,
+			Name:    "updated store",
+		})
+		require.NoError(t, err)
+	})
+
+	t.Run("UpdateStore_with_authz", func(t *testing.T) {
+		openfga := MustNewServerWithOpts(
+			WithDatastore(ds),
+		)
+		t.Cleanup(openfga.Close)
+
+		clientID := "validclientid"
+		settings := newSetupAuthzModelAndTuples(t, openfga, clientID)
+
+		openfga.authorizer = authz.NewAuthorizer(&authz.Config{StoreID: settings.rootData.id, ModelID: settings.rootData.modelID}, openfga, openfga.logger)
+
+		t.Run("error_when_CheckAuthz_errors", func(t *testing.T) {
+			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
+			_, err := openfga.UpdateStore(ctx, &openfgav1.UpdateStoreRequest{
+				StoreId: settings.testData.id,
+				Name:    "updated store",
+			})
+
+			require.ErrorIs(t, err, authz.ErrUnauthorizedResponse)
+		})
+
+		t.Run("successfully_call_UpdateStore", func(t *testing.T) {
+			ctx := authclaims.ContextWithAuthClaims(context.Background(), &authclaims.AuthClaims{ClientID: clientID})
+			settings.addAuthForRelation(ctx, t, authz.CanCallUpdateStore)
+
+			_, err := openfga.UpdateStore(ctx, &openfgav1.UpdateStoreRequest{
+				StoreId: settings.testData.id,
+				Name:    "updated store",
+			})
+
+			require.NoError(t, err)
 		})
 	})
 }